Rename module `built-ins` -> `webapp-builtins`.
