+++ src/main-lactoserv/export/default.js
@@ -1,7 +1,7 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import * as builtIns from '@this/built-ins';
+import * as builtIns from '@this/webapp-builtins';
 import { Host } from '@this/host';
 
 import { Debugging } from '#p/Debugging';
@@ -11,7 +11,7 @@ import { UsualSystem } from '#p/UsualSystem';
 
 export default async function main() {
   // This is just a nominal reference to keep the build system from thinking
-  // that the `built-ins` module is unused. (It _isn't_ used in the framework,
+  // that the `webapp-builtins` module is unused. (It _isn't_ used in the framework,
   // but it still needs to be available when loading configuration files.)
   if (builtIns === null) {
     throw new Error('Something is very wrong.');
