+++ src/main-lactoserv/package.json
@@ -13,7 +13,7 @@
 
   "dependencies": {
     "@this/async": "*",
-    "@this/built-ins": "*",
+    "@this/webapp-builtins": "*",
     "@this/clocks": "*",
     "@this/data-values": "*",
     "@this/host": "*",
