+++ src/sys-framework/README.md
@@ -5,7 +5,7 @@ This module contains the Lactoserv framework, including all the base classes
 and plumbing required for the system.
 
 This module notably does _not_ contain any concrete application or service
-implementations. The built-in ones of those live in `@this/built-ins`.
+implementations. The built-in ones of those live in `@this/webapp-builtins`.
 
 - - - - - - - - - -
 ```
