+++ src/webapp-builtins/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@this/built-ins",
+  "name": "@this/webapp-builtins",
   "version": "0.0.1",
   "type": "module",
   "private": true,
