+++ src/webapp-builtins/private/RateLimitedStream.js
